fix bug in getImage + data type

+++ source/annotatorlib/include/AnnotatorLib/Camera.h
@@ -37,7 +37,7 @@ class ANNOTATORLIB_API Camera : public ImageSet {
    * @param frame
    * @return image
    */
-  virtual Image getImage(unsigned int /*in*/ frame) override;
+  virtual Image getImage(unsigned long /*in*/ frame) override;
 
   virtual bool gotoPosition(unsigned long position) override;
